@@ -15,9 +15,10 @@
     }
   },
   "require": {
-    "php": ">=7.1"
+    "php": ">=7.4"
   },
   "require-dev": {
-    "phpunit/phpunit": "^6.5"
+    "phpunit/phpunit": "^9.0",
+    "squizlabs/php_codesniffer": "3.*"
   }
 }
